+++ app/src/main/java/org/simple/clinic/facility/change/FacilitiesUpdateType.kt
@@ -1,6 +1,0 @@
-package org.simple.clinic.facility.change
-
-enum class FacilitiesUpdateType {
-  FIRST_UPDATE,
-  SUBSEQUENT_UPDATE
-}