+++ app/src/main/java/org/simple/clinic/registration/facility/RegistrationFacilitySelectionUi.kt
@@ -1,6 +1,5 @@
 package org.simple.clinic.registration.facility
 
-import org.simple.clinic.facility.change.FacilitiesUpdateType
 import org.simple.clinic.facility.change.FacilityListItem
 
 interface RegistrationFacilitySelectionUi {
@@ -8,5 +7,5 @@ interface RegistrationFacilitySelectionUi {
   fun hideProgressIndicator()
   fun showToolbarWithSearchField()
   fun showToolbarWithoutSearchField()
-  fun updateFacilities(facilityItems: List<FacilityListItem>, updateType: FacilitiesUpdateType)
+  fun updateFacilities(facilityItems: List<FacilityListItem>)
 }