+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Migrated `RegistrationFacilitySelectionScreen` to Mobius
 - Migrated `RecentPatientsView` to Mobius
 - [In Progress: 16th Jul 2020] Migrate `LoginPinScreen` to Mobius
+- Change selection of facility throughout the codebase to use `ItemAdapter`
 
 ### Changes
 - Updated translations: `bn_IN`, `am_ET`, `te_IN`, `pa_IN`, `hi_IN`, `bn_BD`